+++ include/compression.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "ff.hpp"
+#include "field.hpp"
 #include "ntt.hpp"
 #include "params.hpp"
 #include <cmath>
